Wire up 'alignment' state
Instead of giving every radio button it's own onChange handler and using
the alignment state to set the checked state of the radio buttons on
every render, I've captured the onChange as it bubbles up through a
parent and only set the default checked state on each render. This means
the radio buttons keep their checked state in the regular DOM and we can
only read it back.

The React docs call this an 'uncontrolled component' vs. a 'controlled
component'. The React library yells if you change in input from being
controlled to uncontrolled (or vice versa) during a componenents lifecycle,
but is okay with uncontrolled components otherwise.

The downside to this is that if we try to set the alignment state
anywhere directly (i.e. not where we're just reading it out of the DOM),
the checked state of the buttons will become out of sync. Since we have
no need to do that here, I'm going with this approach because it's a bit
less verbose and it's interesting.